+++ index.html
@@ -44,13 +44,14 @@
             <option value="1">Одна комната</option>
             <option value="2">Две комнаты</option>
             <option value="3">Три комнаты</option>
+            <option value="more">Более трех комнат</option>
           </select>
           <select name="housing-guests" id="housing-guests" class="map__filter">
             <option value="any" selected>Любое число гостей</option>
             <option value="2">Два гостя</option>
             <option value="1">Один гость</option>
-            <option value="0">Не для гостей</option>
-          </select>
+            <option value="more">Более двух гостей</option>
+                 </select>
           <fieldset id="housing-features" class="map__features">
             <input type="checkbox" name="features" value="wifi" id="filter-wifi" class="map__checkbox visually-hidden">
             <label class="map__feature map__feature--wifi" for="filter-wifi">Wi-Fi</label>
@@ -133,7 +134,7 @@
         <fieldset class="ad-form__element">
           <label class="ad-form__label" for="capacity">Количество мест</label>
           <select id="capacity" name="capacity">
-            <option value="3" class ="hidden">для 3 гостей</option>
+                        <option value="3" class ="hidden">для 3 гостей</option>
             <option value="2" class ="hidden">для 2 гостей</option>
             <option value="1" selected>для 1 гостя</option>
             <option value="0" class ="hidden">не для гостей</option>
