+++ js/utils/util.js
@@ -1,7 +1,7 @@
 // Функция взята из интернета и доработана
 // Источник - https://www.freecodecamp.org/news/javascript-debounce-example
 
-function debounce (callback, timeoutDelay = 500) {
+const debounce = (callback, timeoutDelay) => {
   // Используем замыкания, чтобы id таймаута у нас навсегда приклеился
   // к возвращаемой функции с setTimeout, тогда мы его сможем перезаписывать
   let timeoutId;
@@ -13,10 +13,21 @@ function debounce (callback, timeoutDelay = 500) {
 
     // Затем устанавливаем новый таймаут с вызовом колбэка на ту же задержку
     timeoutId = setTimeout(() => callback.apply(this, rest), timeoutDelay);
-
     // Таким образом цикл «поставить таймаут - удалить таймаут» будет выполняться,
     // пока действие совершается чаще, чем переданная задержка timeoutDelay
+    //console.log('done');
   };
-}
+};
 
-export {debounce};
+const showAlert = (message, time) => {
+  const alertMessage = document.createElement('div');
+  alertMessage.classList.add('alert-message');
+  alertMessage.textContent = message;
+  document.body.append(alertMessage);
+
+  setTimeout(() => {
+    alertMessage.remove();
+  }, time);
+};
+
+export {debounce,showAlert};
