+++ js/validation.js
@@ -1,31 +1,14 @@
 const MIN_TITLE_LENGTH = 30;
 const MAX_TITLE_LENGTH = 100;
 const MAX_PRICE = 1000000;
+
+
 const titleInput = document.querySelector('#title');
 const priceInput = document.querySelector('#price');
 const roomsList  = document.querySelector('#room_number');
 const capacityList  = document.querySelector('#capacity');
 const capacityOptions = capacityList.querySelectorAll('option');
-const infoForm = document.querySelector('.ad-form');
-const formFieldsets = infoForm.querySelectorAll('fieldset');
-const mapFiltersForm = document.querySelector('.map__filters');
-const allFilters = mapFiltersForm.querySelectorAll('select');
 
-const makeAllDisabled = () => {
-  infoForm.classList.add('ad-form--disabled');
-  formFieldsets.forEach((fieldset)=> fieldset.setAttribute('disabled','true'));
-  mapFiltersForm.classList.add('ad-form--disabled');
-  allFilters.forEach((filter)=> filter.setAttribute('disabled','true'));
-};
-
-makeAllDisabled();
-
-const makeAllAktive = () => {
-  infoForm.classList.remove('ad-form--disabled');
-  formFieldsets.forEach((fieldset)=> fieldset.removeAttribute('disabled'));
-  mapFiltersForm.classList.remove('ad-form--disabled');
-  allFilters.forEach((filter)=> filter.removeAttribute('disabled'));
-};
 
 titleInput.addEventListener('input', ()=> {
   const titleLength = titleInput.value.length;
@@ -88,4 +71,4 @@ const getRoomGuestsRelation = () => {
 
 roomsList.addEventListener('change', getRoomGuestsRelation);
 
-export {makeAllAktive};
+
